Move auth code into auth view
