+++ scoring_engine/web/views/auth.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template
 
 import bcrypt
-from flask import flash, redirect, request, url_for
-from flask_login import current_user, login_user, logout_user
+from flask import flash, redirect, request, url_for, g
+from flask_login import current_user, login_user, logout_user, LoginManager, current_user
 from flask_wtf import FlaskForm
 
 from wtforms import TextField, PasswordField
@@ -11,7 +11,24 @@ from wtforms.validators import InputRequired
 from scoring_engine.db import db
 from scoring_engine.models.user import User
 
+from scoring_engine.web import app
 mod = Blueprint('auth', __name__)
+
+
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = 'login'
+
+
+@login_manager.user_loader
+def load_user(id):
+    return User.query.get(int(id))
+
+
+@app.before_request
+def get_current_user():
+    g.user = current_user
+
 
 # Creating our login form
 class LoginForm(FlaskForm):
